add fluid.ecl tests
