+++ tests/test_fluids_bw92.py
@@ -1,7 +1,4 @@
-"""Test functions for pem.fluid module
-
-These test functions are designed to test core functionality with pytest
-
+"""Test functions for pem.fluid.bw92 module
 """
 import pathlib
 
@@ -255,24 +252,6 @@ def test_oil_fvf(dummy_values, tol):
     param, ans = dummy_values
     rho, G, rg, t = param
     assert np.allclose(bw92.oil_fvf(rho, G, rg, t), ans, rtol=tol["rel"])
-
-
-@pytest.mark.parametrize(
-    "pres, extrap, ans",
-    [
-        (325, "const", 1.4615),
-        (325, "pchip", 1.4615),
-        (np.r_[325, 375], "const", np.r_[1.4615, 1.4505]),
-        (np.r_[325, 375], "pchip", np.r_[1.4615, 1.4505]),
-    ],
-)
-def test_oil_fvf_table(test_data, pres, ans, extrap, tol):
-    tab = np.loadtxt(test_data / "PVT_BO.inc")
-    assert np.allclose(
-        fluid_ecl.oil_fvf_table(tab[:, 0], tab[:, 1], pres, extrap=extrap),
-        ans,
-        rtol=tol["rel"],
-    )
 
 
 # @pytest.mark.parametrize('')
